# api_server/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: info@payloadops.com
    name: Payload Ops Team
  description: "The Payload Ops API streamlines management of AI prompts, projects,\
    \ organizations, teams, and users through conventional HTTP requests. This platform\
    \ enables efficient automation and control of resources, providing robust tools\
    \ for developers to manage settings, memberships, and activities seamlessly."
  title: Payload Ops API
  version: "1.0"
servers:
- description: Production server (uses live data)
  url: https://api.payloadops.com/v1
- description: Sandbox server (uses test data)
  url: https://sandbox-api.payloadops.com/v1
security:
- ApiKeyAuth: []
- BearerAuth: []
paths:
...
  /organizations/{orgId}/teams:
    get:
      description: |
        Retrieves a comprehensive list of all teams within a specified organization. This endpoint is essential for administrators and organizational leaders to view and manage team structures, understand team compositions, and plan resource allocation effectively.

        **Key Features**:
        - **Comprehensive Overview**: Provides detailed views of all teams, including their names, descriptions, and member counts.
        - **Management and Planning**: Supports strategic planning and management by offering insights into team distributions and roles within the organization.

        **Use Cases**:
        - Executives and managers reviewing team structures for strategic planning.
        - HR departments managing resource allocations and internal audits of team distributions.
      operationId: listTeams
      parameters:
      - description: The identifier of the organization whose teams are being listed.
        explode: false
        in: path
        name: orgId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Team'
                type: array
          description: "Successfully retrieved a list of all teams within the organization,\
            \ each entry containing details such as team ID, name, and member list."
        "500":
          description: "A server error occurred, preventing the retrieval of team\
            \ information."
      security:
      - BearerAuth: []
      summary: List all teams in an organization
      tags:
      - Teams
    post:
      description: |
        Creates a new team within an organization based on the provided details such as team name, description, and intended members. This operation facilitates the expansion and structuring of an organization's teams to meet evolving business needs and project demands.

        **Key Features**:
        - **Team Formation**: Enables the formation of new teams tailored to specific projects or organizational functions.
        - **Customizable Setup**: Allows for detailed definition of team characteristics including leadership roles and member responsibilities.

        **Use Cases**:
        - Forming new project teams as part of business expansion or new project initiation.
        - Restructuring existing departments to enhance efficiency or focus on new strategic areas.
      operationId: createTeam
      parameters:
      - description: The identifier of the organization where the new team will be
          created.
        explode: false
        in: path
        name: orgId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamInput'
        description: "Contains the necessary attributes for creating a new team, including\
          \ name, description, and initial members if applicable."
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: "The team was created successfully and is now part of the organization,\
            \ with detailed information about the new team provided."
        "400":
          description: "Invalid input, typically due to missing required fields or\
            \ improperly formatted data."
        "500":
          description: "A server error occurred, preventing the creation of the new\
            \ team."
      security:
      - BearerAuth: []
      summary: Create a new team for an organization
      tags:
      - Teams
  /organizations/{orgId}/teams/{teamId}:
    delete:
      description: |
        Deletes a specific team from an organization based on its team ID. This operation is critical for managing organizational structure, particularly when a team is dissolved or merged with another.

        **Key Features**:
        - **Streamlined Organizational Structure**: Helps maintain a clean and efficient organizational hierarchy by removing outdated or unnecessary teams.
        - **Security and Compliance**: Ensures that all related data is securely handled and removed in compliance with data protection policies.

        **Use Cases**:
        - Removing teams that are no longer active or have been restructured.
        - Ensuring that team data does not persist in the system unnecessarily, particularly in sensitive environments.
      operationId: deleteTeam
      parameters:
      - description: The identifier of the organization from which the team will be
          deleted.
        explode: false
        in: path
        name: orgId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the team to be deleted.
        explode: false
        in: path
        name: teamId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: "Team deleted successfully, with no remaining data stored."
        "404":
          description: Either the specified team or the organization was not found.
        "500":
          description: "A server error occurred, preventing the deletion of the team."
      security:
      - BearerAuth: []
      summary: Delete a specific team
      tags:
      - Teams
    get:
      description: |
        Retrieves detailed information about a specific team within an organization, using its unique team ID. This operation is crucial for administrators who need to manage team settings, understand team composition, and review team-specific roles and permissions.

        **Key Features**:
        - **Comprehensive Team Details**: Provides complete information about the team, including its members, roles, projects, and operational statistics.
        - **Management and Oversight**: Essential for overseeing team configurations and ensuring that team setups align with organizational policies and goals.

        **Use Cases**:
        - Team leaders reviewing team composition and roles for project assignments.
        - HR and management accessing team details for planning and resource allocation.
      operationId: getTeam
      parameters:
      - description: The identifier of the organization to which the team belongs.
        explode: false
        in: path
        name: orgId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the team to retrieve.
        explode: false
        in: path
        name: teamId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: Successfully retrieved detailed information about the team.
        "404":
          description: The specified team or organization was not found.
        "500":
          description: "A server error occurred, preventing the retrieval of the team."
      security:
      - BearerAuth: []
      summary: Get a specific team by ID
      tags:
      - Teams
    put:
      description: |
        Updates the details of a specific team within an organization. This endpoint allows for the modification of team data such as name, description, and member roles. It supports dynamic team management and ensures that team information remains current and relevant to organizational needs.

        **Key Features**:
        - **Dynamic Team Management**: Enables changes to team structure, roles, and project assignments.
        - **Adaptable to Changing Needs**: Facilitates quick updates to team details in response to project shifts or strategic realignments.

        **Use Cases**:
        - Administrators adjusting team roles and permissions to reflect new project requirements.
        - Team leaders modifying team descriptions or details to better define the team’s purpose or objectives.
      operationId: updateTeam
      parameters:
      - description: The identifier of the organization where the team is registered.
        explode: false
        in: path
        name: orgId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the team to be updated.
        explode: false
        in: path
        name: teamId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamInput'
        description: Contains the updated information for the team.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: "Team information updated successfully, reflecting new settings\
            \ and configurations."
        "400":
          description: "Invalid input, such as missing required data or unsupported\
            \ modifications."
        "404":
          description: Either the specified team or the organization was not found.
        "500":
          description: "A server error occurred, preventing the update of the team."
      security:
      - BearerAuth: []
      summary: Update a specific team
      tags:
      - Teams
...
  /projects/{projectId}/keys/{keyId}:
    delete:
      description: |
        Deletes a specific API key from a project, permanently removing its access to the project resources. This operation is essential for maintaining security when an API key is no longer needed or if its permissions need to be revoked urgently.

        **Key Features**:
        - **Secure Cleanup**: Ensures that unused or compromised API keys are removed to prevent unauthorized access.
        - **Access Revocation**: Provides a method to immediately invalidate an API key when necessary.

        **Use Cases**:
        - Removing access for team members who no longer require it, such as after project completion or role changes.
        - Revoking keys that may have been exposed or compromised to maintain security integrity.
      operationId: deleteApiKey
      parameters:
      - description: The unique identifier of the project from which the API key will
          be deleted.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the API key to be deleted.
        explode: false
        in: path
        name: keyId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: "The API key was deleted successfully, no content returned."
        "404":
          description: Either the API key or the project was not found.
        "500":
          description: "A server error occurred, preventing the deletion of the API\
            \ key."
      security:
      - BearerAuth: []
      summary: Delete a specific API key
      tags:
      - API Keys
    get:
      description: |
        Retrieves detailed information about a specific API key associated with a project. This operation is crucial for managing access to project resources, allowing administrators to review the permissions and details of individual API keys.

        **Key Features**:
        - **Security Management**: Provides the ability to monitor and manage API access through detailed key information.
        - **Access Review**: Enables project managers to verify the scope and validity of an API key, ensuring that it aligns with current security policies.

        **Use Cases**:
        - Administrators auditing access permissions for compliance and security best practices.
        - Developers checking API key details to troubleshoot access issues or to ensure proper configuration.
      operationId: getApiKey
      parameters:
      - description: The unique identifier of the project associated with the API
          key.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the API key to retrieve.
        explode: false
        in: path
        name: keyId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
          description: Detailed information about the API key retrieved successfully.
        "404":
          description: Either the specified API key or the project was not found.
        "500":
          description: "A server error occurred, preventing the retrieval of the API\
            \ key."
      security:
      - BearerAuth: []
      summary: Retrieve a specific API key
      tags:
      - API Keys
    put:
      description: |
        Updates the scopes associated with a specific API key. This operation allows administrators to modify access levels and permissions for API keys, adapting to changing security needs or project requirements.

        **Key Features**:
        - **Scope Management**: Adjusts the access boundaries defined by the API key scopes, ensuring they are kept up-to-date with project roles and permissions.
        - **Flexible Access Control**: Offers the ability to dynamically change API key permissions without needing to recreate keys.

        **Use Cases**:
        - Modifying API key permissions in response to role changes within a project team.
        - Updating key scopes to restrict or expand access as part of security policy adjustments.
      operationId: updateApiKey
      parameters:
      - description: The unique identifier of the project for which the API key is
          managed.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the API key to be updated.
        explode: false
        in: path
        name: keyId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyInput'
        description: JSON payload containing the new scopes for the API key.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
          description: The API key's scopes were updated successfully.
        "400":
          description: "Invalid input, such as unspecified or unsupported scopes."
        "404":
          description: Either the API key or the project was not found.
        "500":
          description: "A server error occurred, preventing the update of the API\
            \ key."
      security:
      - BearerAuth: []
      summary: Update an API key's scopes
      tags:
      - API Keys
...
components:
  schemas:
...
    Team:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        name: name
        description: description
        id: id
        orgId: orgId
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Unique identifier for the team
          type: string
        name:
          description: Name of the team
          type: string
        description:
          description: A brief description of the team
          type: string
        orgId:
          description: The organization to which the team belongs
          type: string
        createdAt:
          description: Timestamp of team creation
          format: date-time
          type: string
        updatedAt:
          description: Timestamp of the last update to the team
          format: date-time
          type: string
      type: object
    TeamInput:
      example:
        name: name
        description: description
      properties:
        name:
          description: Name of the team
          type: string
        description:
          description: A brief description of the team
          type: string
      required:
      - name
      type: object
...
    ApiKey:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        id: id
        scopes:
        - scopes
        - scopes
        projectId: projectId
        key: key
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Unique identifier for the API key
          type: string
        key:
          description: The API key token
          type: string
        scopes:
          description: List of scopes or permissions granted by this API key
          items:
            type: string
          type: array
        projectId:
          description: The project ID this API key belongs to
          type: string
        createdAt:
          description: Timestamp when the API key was created
          format: date-time
          type: string
        updatedAt:
          description: Timestamp when the API key was last updated
          format: date-time
          type: string
      type: object
    ApiKeyInput:
      example:
        scopes:
        - scopes
        - scopes
      properties:
        scopes:
          description: List of scopes or permissions to be granted to this API key
          items:
            type: string
          type: array
      required:
      - scopes
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: |
        API Key authentication provides a straightforward method for clients to authenticate requests. This scheme expects an API key to be supplied in the headers of each request. Clients must include the API key in the 'X-API-KEY' header field.

        **How to use**:
        - Obtain an API key from the API provider.
        - Include the API key in the header of each request to the API as follows:
          ```
          X-API-KEY: {api_key}
          ```
        - The API validates the key against its database, and if the key is valid, allows access to the requested resources.

        **Advantages**:
        - Simple and easy to implement.
        - Suitable for accessing public APIs where security requirements are not excessively strict.

        **Limitations**:
        - Less secure than more complex authentication methods, such as tokens, as API keys can be easily exposed in transit or storage.
        - API keys do not expire by default and revoking them requires direct intervention.
      in: header
      name: X-API-KEY
      type: apiKey
    BearerAuth:
      bearerFormat: JWT
      description: |
        Bearer Token authentication uses JSON Web Tokens (JWT) to verify the identity of the requester. This scheme is included in the Authorization header of each HTTP request using the Bearer format.

        **How to use**:
        - Obtain a JWT from the authentication server. Typically, this involves an initial request with credentials (username/password) to receive the token.
        - Include the token in the Authorization header of your HTTP requests as follows:
          ```
          Authorization: Bearer {token}
          ```
        - The server decodes the JWT to verify its validity and authorizes the request based on the token's payload and signature.

        **Advantages**:
        - Offers more security than API key authentication as the token content is protected by cryptographic algorithms.
        - Tokens can be designed to expire after a certain period, enhancing security by limiting the duration of their validity.

        **Limitations**:
...